refactor: update user rating and reporting components for improved data handling
- Modified userRatingCard and staffRatingCard components to use user_id instead of username for profile selection, enhancing data accuracy.
- Updated report API validation to check for the presence of 'comment' instead of 'reported', ensuring correct error handling.
- Added a new ratings page for user-specific ratings, integrating pagination and improved user feedback display.
- Enhanced layout and messaging across various user pages to provide clearer user experience and visual consistency.
