+++ src/routes/(private)/user/[id]/+page.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-    import type { PageData } from './$types';
+  import type { PageData } from "./$types";
 
-    let { data }: { data: PageData } = $props();
+  let { data }: { data: PageData } = $props();
 </script>
 
-<div class="flex flex-col items-center justify-center min-h-[200px]">
-  <h3 class="text-2xl font-bold mb-4">Overview Coming Soon</h3>
-  <p class="text-gray-600">Stay tuned for more!</p>
+<div class="col-span-full flex flex-col items-center justify-center py-20">
+  <i class="fa-solid fa-info fa-3x mb-4"></i>
+  <h2 class="text-2xl font-semibold mb-2">Overview Coming Soon.</h2>
 </div>
