+++ src/routes/(private)/user/[id]/reviews/+page.server.ts
@@ -1,0 +1,5 @@
+import type { PageServerLoad } from './$types';
+
+export const load = (async () => {
+    return {};
+}) satisfies PageServerLoad;
